file package split into subpackages per format
