--- src/main/java/net/novaware/nes/core/file/NesHeaderWriter.java
@@ -1,4 +1,0 @@
-package net.novaware.nes.core.file;
-
-public class NesHeaderWriter {
-}
