+++ src/main/java/net/novaware/nes/core/file/ines/NesFileHandler.java
@@ -1,4 +1,4 @@
-package net.novaware.nes.core.file;
+package net.novaware.nes.core.file.ines;
 
 public abstract class NesFileHandler {
