+++ src/main/java/net/novaware/nes/core/file/ines/NesHeaderWriter.java
@@ -1,0 +1,4 @@
+package net.novaware.nes.core.file.ines;
+
+public class NesHeaderWriter {
+}
