+++ src/main/java/net/novaware/nes/core/file/nesy/NesyFileReader.java
@@ -1,4 +1,4 @@
-package net.novaware.nes.core.file;
+package net.novaware.nes.core.file.nesy;
 
 public class NesyFileReader {
     // NOTE: support both zipped and exploded
