+++ src/main/java/net/novaware/nes/core/file/nesy/NesyFileWriter.java
@@ -1,4 +1,4 @@
-package net.novaware.nes.core.file;
+package net.novaware.nes.core.file.nesy;
 
 public class NesyFileWriter {
     // NOTE: support both zipped and exploded
