+++ src/main/java/net/novaware/nes/core/file/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Classes related to reading and writing iNES file format
+ * Model of a universal nes file. Format specific handlers in subpackages
  */
 package net.novaware.nes.core.file;
