+++ src/main/java/net/novaware/nes/core/file/unif/UnifFileReader.java
@@ -1,4 +1,4 @@
-package net.novaware.nes.core.file;
+package net.novaware.nes.core.file.unif;
 
 public class UnifFileReader {
     // https://www.nesdev.org/wiki/UNIF
