+++ src/test/groovy/net/novaware/nes/core/file/DataBuilder.groovy
@@ -1,4 +1,6 @@
-package net.novaware.nes.core.file;
+package net.novaware.nes.core.file
+
+import net.novaware.nes.core.file.ines.NesFileHandler;
 
 import java.nio.ByteBuffer
