+++ src/test/groovy/net/novaware/nes/core/file/ines/NesFileWriterSpec.groovy
@@ -1,5 +1,6 @@
-package net.novaware.nes.core.file
+package net.novaware.nes.core.file.ines
 
+import net.novaware.nes.core.file.NesFile
 import net.novaware.nes.core.util.Quantity
 import spock.lang.Specification
