+++ src/test/groovy/net/novaware/nes/core/test/TestDataBuilder.groovy
@@ -1,10 +1,10 @@
-package net.novaware.nes.core.file;
+package net.novaware.nes.core.test;
 
 /**
  * @see <a href="http://www.natpryce.com/articles/000714.html">Test Data Builders: an alternative to the Object Mother pattern</a>
  * @param <T>
  */
-public interface TestDataBuilder<T> {
+interface TestDataBuilder<T> {
 
     T build();
 }
